create order migration

+++ src/Sales.Infrastructure.Product.Data.Dapper/Migration/CreateProductTables.cs
@@ -23,7 +23,7 @@ namespace Sales.Infrastructure.Product.Data.Dapper.Migration
                   .WithColumn("Id").AsInt64().NotNullable().PrimaryKey().Identity()
                   .WithColumn("Title").AsString(150).NotNullable()
                   .WithColumn("CopyNumber").AsInt32().NotNullable()
-                  .WithColumn("Price").AsInt32().NotNullable()
+                  .WithColumn("Price").AsDecimal().NotNullable()
                   .WithColumn("ImagePath").AsString().Nullable()
                   .WithColumn("CreatedDate").AsDateTime().NotNullable()
                   .WithColumn("UpdateDate").AsDateTime().NotNullable();
